[klibc] ipconfig: Do not poll sockets we don't intend to read from
When a device is in state DEVST_COMPLETE or DEVST_ERROR, we never read
packets from its socket, but we still poll it.  If the socket is
readable or in an error state, this results in busy-polling until a
timeout is reached.

Only add a device's socket to the fds array when the device is in some
other state.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>