[klibc] ipconfig: Handle error events from poll
When poll returns events for a socket other than POLLRDNORM, such as
POLLERR, we do not read from the socket but keep polling on it.  This
results in busy-polling while the error state persists.

In case some other event is returned, put the device into DEVST_ERROR
state for 1 second.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>