[klibc] arm: Move shared library below executables
The default base address for arm executables is 64 kiB, and the base
address for the shared library is 3.5 MiB when generating Thumb
instructions.  This limits the size of executables using it to be
< 3.5 MiB, and it also interacts with a bug in QEMU's user-mode
emulation, causing executables to immediately crash.

Moving the shared library to the end of the 16 MiB range didn't help.
Instead, move the shared library to 2 MiB and executables to 4 MiB.
This works for both Thumb and non-Thumb configurations, so make it
unconditional.

References: https://bugs.debian.org/1040981
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>