[klibc] riscv64: Move shared library below executables
The default base address for riscv64 executables is 64 kiB, and I
originally selected a base address for the shared library of 2 MiB.
This limits the size of executables using it to be < 2 MiB, and it
also interacts with a bug in QEMU's user-mode emulation, causing
executables to immediately crash.  It doesn't help code generation,
because single-instruction jumps have a range of only ±1 MiB and
two-instruction jumps have a range of ±2 GiB.

Move the shared library to 64 kiB and executables to 576 kiB, to allow
for generation of single-instruction jumps.  (klibc.so is currently
much smaller than the 512 kiB this allows for.)

References: https://bugs.debian.org/1040981
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>