[klibc] Install command links as appropriate
We build gzip to support being called as "gunzip" or "zcat", and
similarly we build halt to support being called as "poweroff" or
"reboot".  In the build directory, we create the aliases as links
to the same binary.

When installing, however, we pass all the names for the same binary to
the "install" command, creating multiple copies of it in the install
directory.

To fix this:

- Define an install_links command that creates links as specified in
  an install-links-y variable.
- Invoke install_links after installing binaries.
- In usr/gzip/Kbuild and usr/utils/Kbuild, specify the links in
  install-links-y instead of install-y.

Reported-by: Benjamin Drung <benjamin.drung@canonical.com>
References: https://bugs.launchpad.net/ubuntu/+source/klibc/+bug/2028571
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>